+++ packages/releases/src/release-config.ts
@@ -133,7 +133,7 @@ export function distConfigFromReleasesConfig(config: ReleasesConfig): ReleaseDis
     },
     metadata: distMetadata,
     host,
-    release: { generatorVersion: '0.3.1' },
+    release: { generatorVersion: '0.4.0' },
   }
   if (config.contract) out.contract = config.contract
   if (config.filenameTemplate) out.filenameTemplate = config.filenameTemplate